backend/requirements.txt
flask
flask-cors
Pillow
numpy
tensorflow
werkzeug